Moved MarioGAN code into GANProcess
A lot of the MarioGANUtil code is actually not unique to Mario since it
also works with Zelda #291 so the common code was moved to GANProcess.
This required some fixes in various other places.
